// app/ket-luan/types.ts
export interface FormValuesConclude {
  tieu_hoa: "be_hay_tao_bon" | "be_hay_tieu_chay_nhe" | "be_an_uong_kho_tieu" | "be_hap_thu_kem"
  suc_de_khang: "be_hay_bi_om_vat" | "be_mac_benh_duong_ho_hap_man_tinh" | "be_suc_de_khang_kem_hay_met_moi"
  bat_on_ve_giac_ngu: "be_kho_ngu_ngu_khong_sau_giac"
  bat_on_ve_tinh_than: "be_cang_thang_de_cau_gat" | "be_lo_do_thieu_tap_trung"
  bat_on_ve_dinh_duong: "tre_bieng_an" | "tre_an_tot_nhung_hap_thu_kem" | "tre_an_thien_lech" | "tre_co_van_de_tieu_hoa_lau_ngay" | "tre_co_dau_hieu_thieu_vi_chat"
}